requirements.txt
-f /usr/share/pip-wheels
appdirs==1.4.3
bleach==3.1.0
boto3==1.9.130
botocore==1.12.130
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
Django==2.2
django-appconf==1.0.3
django-autotranslate==1.1.0
django-axes==4.5.4
django-cors-headers==2.5.2
django-debug-toolbar==1.11
django-debug-toolbar-template-timings==0.9
django-filter==2.1.0
django-ipware==2.1.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.5
docutils==0.14
ExifRead==2.1.2
futures==3.1.1
geographiclib==1.49
geopy==1.19.0
goslate==1.5.1
html5lib==1.0.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
Markdown==3.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
olefile==0.46
packaging==19.0
Pillow==6.0.0
pipdeptree==0.13.2
polib==1.1.0
PyJWT==1.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.1
webencodings==0.5.1
